+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-11-22
+### Added
+- Added projects virtual environments
+- Logging error while initialization
+- Configuration script for project virtual environment
+### Fixed
+- Fixed get_command_syntax invoked before socket file exist
+### Changed
+- Workers errors in one file
+- Downloading conda in local directory while building project
+
 ## [1.0.0] - 2022-04-25
 ### Added
 - Logger configuration for every execution environment